# Performance improvement to `recalculateBins`