--- a/Tests/SwiftPlotTests/PerformanceTests/performanceTests.swift
+++ b/Tests/SwiftPlotTests/PerformanceTests/performanceTests.swift
@@ -1,0 +1,3 @@
+import XCTest
+
+final class PerformanceTests: SwiftPlotTestCase {}