Around the end of last year I noticed that the time the tests were taking was starting to become a burden, especially when doing quick iterative work. After investigating I saw that a big chunk of it was due to the increase in histogram tests at that time, which were really exercising one specific method which took a big slice of the testing time and is called several times for each test, at least twice per output file. Due to this, I decided to change the algorithm used in `recalculateBins` to speed it up considerably, and consequently the testing time as well.

The algorithms are documented in the code. I tried giving them big O notation in the comments for performance comparison, but I have never done this before, so if it is stupidly wrong please tell me 😜.

Regarding the performance wins when executing all of the tests on my machine, the new tests report this:

before:
<img width="1233" alt="Screenshot 2020-01-27 at 04 27 16" src="https://user-images.githubusercontent.com/278662/73150962-f6660100-40c0-11ea-8565-e997cb3dc0d0.png">
after:
<img width="1233" alt="Screenshot 2020-01-27 at 04 28 25" src="https://user-images.githubusercontent.com/278662/73150964-f8c85b00-40c0-11ea-8a5d-00dee7995314.png">
they are two tests testing two different algorithms, one used when the data is sorted and another for when its not;

and when executing all the tests:

|                | Without New Tests | With New Tests |
|----------------|---------------|------------|
| master         |    ~ 10.1 s   |  ~ 13.7 s  |
| new algorithms |    ~ 7.3 s    |   ~ 8.1 s  |

The new tests added by this PR add some time to the overall testing time, but it is still an improvement on how it was before, even without the new tests.

Now the time `recalculateBins` takes is basically negligible.

At the time the improvements in testing speed seemed better because there are more tests now and it looks less impressive 😜.

Please tell me about any issues you encounter. I only wanted to test that specific method, but it was marked `private` so I created another internal method that just calls it for testing purposes. I don't know if there is a better approach other than measuring the performance of that private method indirectly by measuring the whole rendering process, or just by changing the access of that method back to `public` like it was before. @KarthikRIyer @karwa.