@KarthikRIyer 2020-01-27
I like this approach! I hadn't paid much attention to the time complexity of this function although I knew it wasn't good. I tried to improve it once but messed it up and forgot about it later 😅.

I'm confused about one thing though. Why do you set the isSorted flag to true when the data is added using the function addSeries()? The user can surely pass in unsorted data, right?