@odmir 2020-01-27
Ok I've changed the `class` back into a `struct`.
Regarding the O notation in the sorted case it should become O(n) and in the unsorted case it should be O(n log(m)) as well assuming n >> m right?
If we assume that in general m ~= sqrt(n) then it should be something like: O(n) if sorted; O(n log(√n)) = O(n log(n)/2 ) if unsorted; right?