@odmir 2020-01-29
> Unfortunately, XCTest performance tests are kind of useless for CI, because we can't set the baseline measurements. I would really like them, but there's not much point to adding them at this time :(

@karwa is there some solution that you know of? Also is there any way to disable some tests, or separate them into another "module" so that CI only runs the non-performance related tests for now?