@odmir 2020-01-29
Ok the build seems to be failing because some other behaviour is dependent on the data being sorted, it would have broken sooner or later, at least as soon as someone tried to plot a histogram after manually changing data.