@odmir 2020-01-30
I had to put a workaround in `layoutData` because code over there was getting the max and min X by getting the first and last elements from `data` causing trouble later on. I added a comment documenting the workaround and the why as well as a `FIXME` because we are not handling when the user uses an empty array, and a "performance" `TODO` because now we are going through the entire `data`, however many series the histogram has.