@odmir 2020-01-30
Yes it is still better. For the benchmarks I created:
1. one branch from a commit where we had: sorting, faster algorithm and fast maxX, minX;
2. another branch from latest commit with: no sorting, fast algorithm and slower maxX, minX.

From both branches I removed the performance tests because they only exercised the algorithm and sorting parts but never touched the maxX and minX.

Then I measured only the `HistogramTests` test suite:

Branch|Average Time (s)
:-:|:-:
1|1.648
2|1.567

so we are better now with an improvement of about 80 ms.