@odmir 2020-01-30
and compared to master:

Branch|Average Time (s)
:-:|:-:
2|1.567
master|4.323

an improvement of 2.756 s.