@karwa 2020-01-30
Could you maybe move the test to a new test suite? Maybe `PerformanceTests`? Then we could add performance tests for other plots there, and you can easily exclude them all if you don't want to run them.